blue_market/requirements.txt
Django===5.1.2
pillow===11.0.0
django-crispy-forms===2.3
django-phonenumber-field===8.0.0
crispy-bootstrap4===2024.10
django-import-export==4.2.1


